scripts/requirements.txt
numpy==1.19.5
PyYAML==5.4.1
